This sort of works, but not happy about dependency.
The remote-file needs to know the name of the class it is registered as
in order to call the correct `send()` method. Not sure how to get past
this.

Maybe set an attribute on every ZjsComponent newClass that is created,
so that the `ZjsComponent.send()` method can call
`closest('[zjs-component]')`? This is going to break nesting, though.

Maybe create a new static method `ZjsComponent.sendTo('tag-name',
objOrId, method, ...args)`?